GitHub Actions: Avoid duplicate tag names
Because if you execute GitHub Actions with another commit that has the same first 7 characters of the commit ID, the tag name will be duplicated.

Append a number (from 1) to the end of the tag name if the exact commit ID matches (This is fairly rare and only happens when the SHA-1 of the commit is duplicated).